Bug fix: Check version of OpenMEEG on MacOS

+++ toolbox/forward/bst_openmeeg.m
@@ -108,7 +108,7 @@ if strcmpi(osType, 'mac64') && ~isdir(bst_fullfile(OpenmeegDir, 'lib'))
 end
 
 % If binary file doesnt exist: download
-if ~isdir(OpenmeegDir) || isempty(dir(bst_fullfile(OpenmeegDir, 'om_gain*'))) || ~strcmpi(prevUrl, url) || isUpdate
+if ~isdir(OpenmeegDir) || (~strcmpi(osType, 'mac64') && isempty(dir(bst_fullfile(OpenmeegDir, 'om_gain*')))) || ~strcmpi(prevUrl, url) || isUpdate
     % If folder exists: delete
     if isdir(OpenmeegDir)
         file_delete(OpenmeegDir, 1, 3);
